Reworked the file structure and lots of testing
Took longer than expected, but the java app is working with the resources directory moved to outside the java app classpath (maven looks for it at {project}/src/main/resources and I moved it to {project}/../resources). Note that depending on where the app is run, the resources won't be found: this was the thing I spent most of my time trying to fix, but I may have to move on for now.